+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `formatForMobileDialing`
 - Added `extension` to GraphQL type
 - Added `format` to GraphQL type
+- Added `formatForCountry` to GraphQL type
 - Added `timezones` to GraphQL type
 - Added support for `locale` and `region` arguments to `description` GraphQL field